requirements.txt
Django>=3.0
mysqlclient>=2.0.0
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
ipaddress